requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-admin-view-permission==1.9
django-appconf==1.0.4
django-extensions==3.1.2
django-froala-editor==3.2.2
django-image-cropping==1.5.0
django-imagefit==0.6.2
django-request-cache==1.2
django-user-agents==0.4.0
django-userforeignkey==0.3.0
django-uuid-upload-path==1.0.0
djangorestframework==3.12.4
easy-thumbnails==2.7.1
idna==2.10
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
ua-parser==0.10.0
urllib3==1.26.4
user-agents==2.2.0
